HZN-1315: Refactor the AWS SQS Sink impl. into different modules, allowing the client (producer) and server (consumer) to be loaded separately.
