--- core/ipc/sink/aws-sqs-impl/src/test/java/org/opennms/core/ipc/sink/aws/sqs/heartbeat/HeartbeatModule.java
@@ -1,86 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2017 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2017 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.core.ipc.sink.aws.sqs.heartbeat;
-
-import org.opennms.core.ipc.sink.api.AggregationPolicy;
-import org.opennms.core.ipc.sink.api.AsyncPolicy;
-import org.opennms.core.ipc.sink.xml.AbstractXmlSinkModule;
-
-/**
- * The Class HeartbeatModule.
- * 
- * @author <a href="mailto:agalue@opennms.org">Alejandro Galue</a>
- */
-public class HeartbeatModule extends AbstractXmlSinkModule<Heartbeat,Heartbeat> {
-
-    /** The Constant INSTANCE. */
-    public static final HeartbeatModule INSTANCE = new HeartbeatModule();
-
-    /**
-     * Instantiates a new heartbeat module.
-     */
-    public HeartbeatModule() {
-        super(Heartbeat.class);
-    }
-
-    /* (non-Javadoc)
-     * @see org.opennms.core.ipc.sink.api.SinkModule#getNumConsumerThreads()
-     */
-    @Override
-    public int getNumConsumerThreads() {
-        return 1;
-    }
-
-    /* (non-Javadoc)
-     * @see org.opennms.core.ipc.sink.api.SinkModule#getId()
-     */
-    @Override
-    public String getId() {
-        return "Heartbeat";
-    }
-
-    /* (non-Javadoc)
-     * @see org.opennms.core.ipc.sink.api.SinkModule#getAggregationPolicy()
-     */
-    @Override
-    public AggregationPolicy<Heartbeat, Heartbeat, Heartbeat> getAggregationPolicy() {
-        // No aggregation
-        return null;
-    }
-
-    /* (non-Javadoc)
-     * @see org.opennms.core.ipc.sink.api.SinkModule#getAsyncPolicy()
-     */
-    @Override
-    public AsyncPolicy getAsyncPolicy() {
-        // Only synchronous dispatching
-        return null;
-    }
-
-}
