+++ core/ipc/sink/aws-sqs/client/src/main/resources/OSGI-INF/blueprint/blueprint-ipc-client.xml
@@ -1,0 +1,27 @@
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0"
+    xmlns:ext="http://aries.apache.org/blueprint/xmlns/blueprint-ext/v1.5.0"
+    xsi:schemaLocation="
+		http://www.osgi.org/xmlns/blueprint/v1.0.0 
+		http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+		http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0
+		http://aries.apache.org/schemas/blueprint-cm/blueprint-cm-1.3.0.xsd
+		http://aries.apache.org/blueprint/xmlns/blueprint-ext/v1.5.0
+		http://aries.apache.org/schemas/blueprint-ext/blueprint-ext-1.5.xsd
+">
+
+	<reference id="awsSqsConfig" interface="org.opennms.core.ipc.common.aws.sqs.AmazonSQSConfig" />
+
+	<bean id="awsSqsManager" class="org.opennms.core.ipc.common.aws.sqs.DefaultAmazonSQSManager">
+		<argument ref="awsSqsConfig"/>
+	</bean>
+
+	<bean id="awsRemoteMessageDispatcherFactory" class="org.opennms.core.ipc.sink.aws.sqs.client.AmazonSQSRemoteMessageDispatcherFactory"
+	  init-method="init" destroy-method="destroy">
+	  <property name="awsSqsManager" ref="awsSqsManager"/>
+	</bean>
+
+    <service ref="awsRemoteMessageDispatcherFactory" interface="org.opennms.core.ipc.sink.api.MessageDispatcherFactory"/>
+
+</blueprint>
