+++ core/ipc/sink/aws-sqs/server/src/main/resources/META-INF/opennms/applicationContext-ipc-sink-aws-server.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xmlns:context="http://www.springframework.org/schema/context"
+  xmlns:util="http://www.springframework.org/schema/util"
+  xmlns:onmsgi="http://xmlns.opennms.org/xsd/spring/onms-osgi"
+  xsi:schemaLocation="
+  http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.2.xsd
+  http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.2.xsd
+  http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util-4.2.xsd
+  http://xmlns.opennms.org/xsd/spring/onms-osgi http://xmlns.opennms.org/xsd/spring/onms-osgi.xsd
+">
+
+  <context:annotation-config />
+
+  <bean id="awqSqsConfig" class="org.opennms.core.ipc.common.aws.sqs.MapBasedSQSConfig" />
+
+  <bean id="awsSqsManager" class="org.opennms.core.ipc.common.aws.sqs.DefaultAmazonSQSManager">
+    <constructor-arg ref="awqSqsConfig"/>
+  </bean>
+
+  <bean id="awsMessageConsumerManager" class="org.opennms.core.ipc.sink.aws.sqs.server.AmazonSQSMessageConsumerManager">
+    <property name="awsSqsManager" ref="awsSqsManager"/>
+  </bean>
+  <onmsgi:service ref="awsMessageConsumerManager" interface="org.opennms.core.ipc.sink.api.MessageConsumerManager" />
+</beans>
