+++ core/ipc/sink/aws-sqs/server/src/main/resources/META-INF/opennms/component-dao.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans" 
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xmlns:context="http://www.springframework.org/schema/context"
+       xsi:schemaLocation="
+       http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
+       http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.2.xsd
+       ">
+
+    <!-- Conditionally load the AWS SQS Sink -->
+    <bean class="org.opennms.core.ipc.sink.aws.sqs.server.AmazonSQSConditionalSinkContext"/>
+
+</beans>
