+++ core/ipc/sink/aws-sqs/server/src/main/resources/OSGI-INF/blueprint/blueprint-ipc-server.xml
@@ -1,0 +1,23 @@
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0"
+           xsi:schemaLocation="
+		http://www.osgi.org/xmlns/blueprint/v1.0.0
+		https://osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+		http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0
+		http://aries.apache.org/schemas/blueprint-cm/blueprint-cm-1.3.0.xsd
+		http://camel.apache.org/schema/blueprint
+		http://camel.apache.org/schema/blueprint/camel-blueprint-2.19.1.xsd
+">
+
+    <reference id="awsSqsConfig" interface="org.opennms.core.ipc.common.aws.sqs.AmazonSQSConfig" />
+    <bean id="awsSqsManager" class="org.opennms.core.ipc.sink.aws.sqs.DefaultAmazonSQSManager">
+        <argument ref="awsSqsConfig"/>
+    </bean>
+
+    <bean id="awsSqsMessageconsumerManager" class="org.opennms.core.ipc.sink.aws.sqs.server.AmazonSQSMessageConsumerManager">
+        <property name="awsSqsManager" ref="awsSqsManager"/>
+    </bean>
+    <service ref="awsMessageConsumerManager" interface="org.opennms.core.ipc.sink.api.MessageConsumerManager" />
+
+</blueprint>
