+++ features/minion/repository/pom.xml
@@ -257,8 +257,8 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.opennms.core.ipc.sink</groupId>
-            <artifactId>org.opennms.core.ipc.sink.aws-sqs-impl</artifactId>
+            <groupId>org.opennms.core.ipc.sink.aws.sqs</groupId>
+            <artifactId>org.opennms.core.ipc.sink.aws.sqs.client</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
